Boa pessoal.

+++ src/main/java/com/mobisi/api/dto/CreateUserDto.java
@@ -45,5 +45,5 @@ public class CreateUserDto {
     private String state;
 
     @NotNull(message = "Invalid disabilityType: disabilityType is NULL")
-    private Integer disabilityType;
+    private Long disabilityType;
 }
